Jogosultságok ellenőrzése
Addresses:
      - createAddress (csak adatbázisban van használva ezért ki lett kommentelve, később akár törölve is lesz)
      - updateAddressById
      - deleteAddressById

+++ Backend/SzakemberKereso/src/main/java/szakemberkereso/Controller/AddressesController.java
@@ -4,6 +4,7 @@
  */
 package szakemberkereso.Controller;
 
+import java.util.Objects;
 import javax.ejb.Stateless;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriInfo;
@@ -16,7 +17,9 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import szakemberkereso.Configuration.Roles;
 import szakemberkereso.Model.Addresses;
+import szakemberkereso.Model.Users;
 import szakemberkereso.Service.AddressesService;
 import szakemberkereso.Service.AuthService;
 
@@ -69,31 +72,46 @@ public class AddressesController {
         return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
     }
     
-    @POST
-    @Path("createAddress")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response createAddress(Addresses address){
-        String result = as.createAddress(address);
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
-    }
+//    @POST
+//    @Path("createAddress")
+//    @Consumes(MediaType.APPLICATION_JSON)
+//    public Response createAddress(Addresses address){
+//        String result = as.createAddress(address);
+//        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+//    }
     
     @POST
     @Path("updateAddressById")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response updateAddressById(Addresses address){
-        String result = as.updateAddressById(address);
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (authService.isUserAuthorized(address.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.USER, Roles.WORKER})) {
+            if(!authService.isUserAuthorized(address.getCurrentUserId(), new Roles[]{Roles.ADMIN})){
+                //a nem ADMIN jogosultságú user-ek csak a saját address-üket módosíthatják
+                if(!Objects.equals(address.getId(), Users.getUserById(address.getCurrentUserId()).getAddressId())){
+                    return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+                }
+            }
+            String result = as.updateAddressById(address);
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        return Response.status(Response.Status.UNAUTHORIZED).entity("Nem sikerült azonosítani.").build();
     }
     
     @POST
     @Path("deleteAddressById")
     @Consumes(MediaType.APPLICATION_JSON)
     public Response deleteAddressById(Addresses address){
-//        if (!authService.isUserAuthorized(address.getUserId(), "ADMIN")) {
-//            return Response.status(Response.Status.FORBIDDEN).build();
-//        }
-        Boolean result = as.deleteAddressById(address.getId());
-        return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        if (authService.isUserAuthorized(address.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.USER, Roles.WORKER})) {
+            if(!authService.isUserAuthorized(address.getCurrentUserId(), new Roles[]{Roles.ADMIN})){
+                //a nem ADMIN jogosultságú user-ek csak a saját address-üket törölhetik
+                if(!Objects.equals(address.getId(), Users.getUserById(address.getCurrentUserId()).getAddressId())){
+                    return Response.status(Response.Status.FORBIDDEN).entity("Nincs jogosultsága ehhez a kéréshez.").build();
+                }
+            }
+            Boolean result = as.deleteAddressById(address.getId());
+            return Response.status(Response.Status.OK).entity(result).type(MediaType.APPLICATION_JSON).build();
+        }
+        return Response.status(Response.Status.UNAUTHORIZED).entity("Nem sikerült azonosítani.").build();
     }
     
 }
